+++ app/index.js
@@ -1,6 +1,6 @@
 /* eslint-disable no-underscore-dangle */
 // yeoman expects underscore dangles for private methods
-const yeoman = require('yeoman-generator');
+const Generator = require('yeoman-generator');
 const spawn = require('cross-spawn');
 const glob = require('glob');
 const fs = require('fs-extra');
@@ -12,12 +12,12 @@ const chalk = require('chalk');
 const commandExists = require('command-exists');
 const findParentDir = require('find-parent-dir');
 const generatorPackageJson = require('../package.json');
-const pascalcase = require('pascalcase');
+const pascalcase = require('../util/pascalcase');
 const { FILE_DELIM_OPEN, FILE_DELIM_CLOSE } = require('../util/ejs-util');
 
 const generatorVersion = generatorPackageJson.version;
 
-module.exports = yeoman.extend({
+module.exports = class extends Generator {
   _customAppName() {
     const privateNpmRe = /@.*\/(.*)/;
     let appname = this._getPackageProp('name');
@@ -234,4 +234,4 @@ Get started:
       `);
     }
   },
-});
+};